# openapi.yaml
openapi: 3.0.3
info:
  title: LeadRecovery API
  description: >
    Multi-tenant lead recovery and reactivation API. Every route except
    /health, /ready, and the provider webhooks requires
    `Authorization: Bearer <tenant api key>`; admin routes require
    `Authorization: Bearer <ADMIN_API_KEY>` instead. See README.md for the
    full narrative reference — this file is the machine-readable version of
    the same routes.
  version: "0.2.0"
servers:
  - url: http://localhost:3000
    description: Local demo server (npm run dev)
...
security: []
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: SaaS-style sign-in
      description: >
        Resolves email+password to the tenant's real API key rather than
        minting a session of its own — every dashboard already knows how
        to store/send an API key. Returns the same generic 401 whether the
        email doesn't match a tenant, that tenant never set a password, or
        the password is wrong, so a caller can't enumerate which emails
        have an account.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                type: object
                properties:
                  apiKey: { type: string }
                  tenant: { $ref: "#/components/schemas/PublicTenant" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Suspended" }
  /auth/forgot-password:
    post:
      tags: [auth]
      summary: Request a password-reset email
      description: >
        Always the same generic response whether or not the email matches
        a tenant — see /auth/login's own note on why. Emails a link (via
        PLATFORM_SENDGRID_API_KEY, or logged to the console if unset) that
        POST /auth/reset-password's token is good for 24 hours.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string }
      responses:
        "200":
          description: Generic success — check your email
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  message: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
  /auth/reset-password:
    post:
      tags: [auth]
      summary: Set a new password from a reset/setup token
      description: >
        The same endpoint sets a tenant's *first* password (the link a
        newly-set `email` gets at creation — see POST /admin/tenants) and a
        later password reset — mechanically identical either way. The
        token is single-use and expires after 24 hours.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string }
                newPassword: { type: string, minLength: 8 }
      responses:
        "200":
          description: Password set
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  apiKey: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
  /auth/request-code:
    post:
      tags: [auth]
      summary: "Phone-number sign-in, step 1: request a one-time code"
      description: >
        The primary sign-in path going forward — no password to set or
        forget. Always the same generic response whether or not the phone
        matches a tenant's loginPhone — see /auth/login's own note on why.
        Texts a 6-digit code (via PLATFORM_SMS_FROM_NUMBER, or logged to the
        console if unset) that POST /auth/verify-code's code is good for 10
        minutes. The code itself is never returned here, unlike the
        password-setup link POST /admin/tenants can return — this endpoint
        takes only a phone number, so echoing the code back would let
        anyone with a phone number log in as that tenant.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string, description: "E.164, e.g. +15551234567." }
      responses:
        "200":
          description: Generic success — check your phone
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  message: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
  /auth/verify-code:
    post:
      tags: [auth]
      summary: "Phone-number sign-in, step 2: verify the code"
      description: >
        Resolves phone+code to the tenant's real API key, same pattern as
        /auth/login. Locks out after 5 wrong attempts since the last
        requested code until a fresh one is requested (a 6-digit code has
        far less entropy than a password). The same generic 401 covers an
        unknown phone, an expired code, a locked-out attempt count, and a
        wrong code.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code]
              properties:
                phone: { type: string }
                code: { type: string }
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                type: object
                properties:
                  apiKey: { type: string }
                  tenant: { $ref: "#/components/schemas/PublicTenant" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Suspended" }
...
components:
  securitySchemes:
    tenantAuth:
      type: http
      scheme: bearer
      description: A tenant's own API key.
    adminAuth:
      type: http
      scheme: bearer
      description: ADMIN_API_KEY.
...
  responses:
    Unauthorized:
      description: Missing or invalid Authorization header
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Suspended:
      description: Tenant is suspended
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
    QuietHours:
      type: object
      properties:
        startHour: { type: integer, minimum: 0, maximum: 23 }
        endHour: { type: integer, minimum: 0, maximum: 23 }
    MessageTemplates:
      type: object
      properties:
        initialGrounded: { type: string }
        initialUngrounded: { type: string }
        followUps: { type: array, items: { type: string } }
        notInterestedCloser: { type: string }
        appointmentReminder:
          type: string
          description: "Sent ~24 hours before a booked appointment. Supports {name}/{businessName}/{appointmentTime}."
        winBack:
          type: string
          description: "Periodic check-in to an opted-in past customer. Supports {name}/{businessName}/{serviceClause}."
...
    PublicTenant:
      type: object
      description: Tenant shape with all secrets stripped (no apiKey, no raw channel credentials).
      properties:
        id: { type: string }
        name: { type: string }
        timezone: { type: string }
        quietHours: { $ref: "#/components/schemas/QuietHours" }
        devMode: { type: boolean }
        channels:
          type: object
          properties:
            sms: { type: boolean }
            whatsapp: { type: boolean }
            email: { type: boolean }
        notifyWebhookUrl: { type: string }
        templates: { $ref: "#/components/schemas/MessageTemplates" }
        knowledgeBase: { type: string }
        autoReplyEnabled: { type: boolean }
        status: { type: string, enum: [active, suspended] }
        statusReason:
          type: string
          enum: [manual, billing]
          description: Why `status` last changed — an admin's own PATCH ("manual") vs. an automatic /webhooks/paddle change ("billing"). Informational only.
        paddleSubscriptionId: { type: string, description: "Paddle subscription id (sub_...) backing this tenant's billing, if any." }
        paddleLastEventAt: { type: string, description: "occurred_at of the most recent Paddle webhook event applied to this tenant's status, as Paddle sent it." }
        contactPhone: { type: string, description: "Reference-only business contact phone, if set at onboarding." }
        contactEmail: { type: string, description: "Reference-only business contact email, if set at onboarding." }
        website: { type: string, description: "Reference-only business website, if set at onboarding." }
        termsAcceptedAt: { type: string, format: date-time, description: "When this tenant accepted the Terms of Service/Privacy Policy. Unset blocks all outbound sending." }
        termsVersion: { type: string, description: "Which CURRENT_TERMS_VERSION was accepted, or \"grandfathered\" for a tenant that predates this field." }
        consentBasisConfirmedAt: { type: string, format: date-time, description: "When the agency confirmed this client has a lawful basis to contact its leads. An audit record, not enforced server-side." }
        termsAttestedAt: { type: string, format: date-time, description: "When the agency recorded that this client already agreed to the Terms of Service/Privacy Policy outside the app. An onboarding record only — does not substitute for termsAcceptedAt." }
        carrierApprovalConfirmedAt: { type: string, format: date-time, description: "When the agency confirmed 10DLC/WhatsApp carrier approval. Unset blocks the sms/whatsapp channels (falls back to email, or skips the lead)." }
        botDisclosureEnabled: { type: boolean, description: "Whether the chatbot's first auto-reply per conversation proactively discloses it's automated. Defaults to true when unset." }
        dataRetentionDays: { type: integer, description: "Days after a lead closes out before the worker auto-purges it. Defaults to 365 when unset." }
        email: { type: string, description: "The tenant's login identity for /auth/login, if set." }
        loginPhone: { type: string, description: "The tenant's phone-based login identity for /auth/request-code + /auth/verify-code, if set — the primary sign-in path going forward." }
        hasPassword: { type: boolean, description: "Whether this tenant has ever set a password — settings.html uses this to show \"Set password\" vs \"Change password\"." }
        publicFormKey: { type: string, description: "Low-privilege token for POST /public/leads/{tenantId} — safe to publish in the tenant's own website (see public/settings.html's embed snippet)." }
        numberHostingOrder: { $ref: "#/components/schemas/NumberHostingOrder" }
        winBackEnabled: { type: boolean, description: "Opt-in: periodic win-back check-ins for converted, opted-in leads. Off by default." }
        winBackCooldownDays: { type: integer, description: "Days between win-back check-ins for one lead. Defaults to 180 when unset." }
        attentionSlaHours: { type: integer, description: "Opt-in: hours before a stale needs-attention item fires a one-time operator alert. Unset disables it." }
        createdAt: { type: string, format: date-time }
...
    NumberHostingOrder:
      type: object
      description: >
        An in-progress or completed "connect this client's existing number"
        request (see src/numberHosting.ts) — the missedcall.io-style
        alternative to assigning a brand new number.
      properties:
        orderSid: { type: string }
        phoneNumber: { type: string }
        status:
          type: string
          description: Mirrors Twilio's own HostedNumberOrder status values verbatim.
          enum: [received, pending-verification, verified, pending-loa, carrier-processing, testing, completed, action-required, failed]
        nextStep: { type: string, description: "Twilio's own description of what happens next." }
        failureReason: { type: string, description: 'Only set once status is "action-required" or "failed".' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...